requirements.txt
openai
pdfplumber
python-docx